+++ scripts/e2e.py
@@ -3,7 +3,7 @@ from string import hexdigits
 import argparse
 from functools import partial
 
-from kinto_client import Client, exceptions as kinto_exceptions
+from kinto_http import Client, exceptions as kinto_exceptions
 from kinto_signer.serializer import canonical_json
 from kinto_signer.hasher import compute_hash
 from kinto_signer.signer.local_ecdsa import ECDSASigner