+++ tests/test_plugin_setup.py
@@ -7,9 +7,10 @@ from pyramid import testing
 from pyramid.exceptions import ConfigurationError
 from requests import exceptions as requests_exceptions
 
-from kinto_signer import on_collection_changed, __version__ as signer_version
+from kinto_signer import __version__ as signer_version
 from kinto_signer.signer.autograph import AutographSigner
 from kinto_signer import includeme
+from kinto_signer.listeners import sign_collection_data
 from kinto_signer import utils
 
 from .support import BaseWebTest, get_user_headers
@@ -26,6 +27,10 @@ class HelloViewTest(BaseWebTest, unittest.TestCase):
             "description": "Digital signatures for integrity and authenticity of records.",  # NOQA
             "url": ("https://github.com/Kinto/kinto-signer#kinto-signer"),
             "version": signer_version,
+            "to_review_enabled": False,
+            "group_check_enabled": False,
+            "editors_group": "editors",
+            "reviewers_group": "reviewers",
             "resources": [
                 {"destination": {"bucket": "alice",
                                  "collection": "destination"},
@@ -123,21 +128,21 @@ class IncludeMeTest(unittest.TestCase):
 class OnCollectionChangedTest(unittest.TestCase):
 
     def setUp(self):
-        patch = mock.patch('kinto_signer.LocalUpdater')
+        patch = mock.patch('kinto_signer.listeners.LocalUpdater')
         self.updater_mocked = patch.start()
         self.addCleanup(patch.stop)
 
     def test_nothing_happens_when_resource_is_not_configured(self):
         evt = mock.MagicMock(payload={"bucket_id": "a", "collection_id": "b"})
-        on_collection_changed(evt, resources=utils.parse_resources("c/d;e/f"))
+        sign_collection_data(evt, resources=utils.parse_resources("c/d;e/f"))
         assert not self.updater_mocked.called
 
     def test_nothing_happens_when_status_is_not_to_sign(self):
         evt = mock.MagicMock(payload={"bucket_id": "a", "collection_id": "b"},
                              impacted_records=[{
                                  "new": {"id": "b", "status": "signed"}}])
-        on_collection_changed(evt, resources=utils.parse_resources("a/b;c/d"))
-        assert not self.updater_mocked.called
+        sign_collection_data(evt, resources=utils.parse_resources("a/b;c/d"))
+        assert not self.updater_mocked.sign_and_update_destination.called
 
     def test_updater_is_called_when_resource_and_status_matches(self):
         evt = mock.MagicMock(payload={"bucket_id": "a", "collection_id": "b"},
@@ -148,7 +153,8 @@ class OnCollectionChangedTest(unittest.TestCase):
         evt.request.registry.signers = {
             "/buckets/a/collections/b": mock.sentinel.signer
         }
-        on_collection_changed(evt, resources=utils.parse_resources("a/b;c/d"))
+        evt.request.route_path.return_value = "/v1/buckets/a/collections/b"
+        sign_collection_data(evt, resources=utils.parse_resources("a/b;c/d"))
         self.updater_mocked.assert_called_with(
             signer=mock.sentinel.signer,
             storage=mock.sentinel.storage,
@@ -162,7 +168,7 @@ class OnCollectionChangedTest(unittest.TestCase):
     def test_updater_does_not_fail_when_payload_is_inconsistent(self):
         # This happens with events on default bucket for kinto < 3.3
         evt = mock.MagicMock(payload={"subpath": "collections/boom"})
-        on_collection_changed(evt, resources=utils.parse_resources("a/b;c/d"))
+        sign_collection_data(evt, resources=utils.parse_resources("a/b;c/d"))
 
 
 class BatchTest(BaseWebTest, unittest.TestCase):
@@ -241,7 +247,7 @@ class SigningErrorTest(BaseWebTest, unittest.TestCase):
         rc = '/buckets/alice/collections/source'
         self.app.app.registry.signers[rc].server_url = 'http://0.0.0.0:1234'
 
-        self.app.put_json("/buckets/alice/collections/source",
-                          {"data": {"status": "to-sign"}},
-                          headers=headers,
-                          status=503)
+        self.app.patch_json("/buckets/alice/collections/source",
+                            {"data": {"status": "to-sign"}},
+                            headers=headers,
+                            status=503)